+++ src/main/java/roomescape/persistence/H2ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.persistence;
 
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.jdbc.core.JdbcTemplate;
